[#2919] Kernel doesn't boot with large fs
git-svn-id: svn://localhost/svn/u-boot/trunk@697 e8ad1c53-384c-47db-b5ce-66ae1f620dea

+++ u-boot-1.1.6/common/cmd_elf.c
@@ -285,7 +285,7 @@ unsigned long load_elf_image (unsigned long addr)
 	unsigned char *image;		/* Binary image pointer             */
 	int i, j;			/* Loop counter                     */
 	unsigned char match;
-	unsigned int entry;
+	unsigned long entry;
 
 	/* -------------------------------------------------- */
